backend/pyproject.toml
[project]
name = "mcp-registry-backend"
version = "0.1.0"
description = "FastAPI backend for MCP Registry with PostgreSQL database and Entra ID authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyjwt[crypto]>=2.9.0",
    "python-dotenv>=1.0.0",
    "python-json-logger>=3.1.0",
    "email-validator>=2.1.0",
]
